Rename folder/file name to hypen case
