+++ src/components/table/table-container/index.ts
@@ -1,0 +1,2 @@
+export * from './styles'
+export * from './table-container'
