+++ src/components/table/table-container/styles.ts
@@ -1,0 +1,3 @@
+import { styled } from '@linaria/react'
+
+export const ElTableContainer = styled.div``
