+++ src/components/table/table-container/table-container.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta } from '@storybook/react'
 // import { figmaDesignUrls } from '../../storybook/figma'
-import { TableContainer } from './tableContainer'
-import { TableToolbar } from '../tableToolbar'
+import { TableContainer } from './table-container'
+import { TableToolbar } from '../table-toolbar'
 import { Menu, MenuItem, MenuItemGroup, MenuList } from '../../menu'
 import { MenuPopover, MenuTrigger } from '#src/components/menu/menu-popover'
 import { Button } from '../../button'
