--- src/components/table/tableContainer/index.ts
@@ -1,2 +1,0 @@
-export * from './styles'
-export * from './tableContainer'
