--- src/components/table/tableContainer/styles.ts
@@ -1,6 +1,0 @@
-import { styled } from '@linaria/react'
-
-export const ElTableContainer = styled.div`
-  gap: 0px;
-  background: var(--fill-white);
-`
