+++ src/index.ts
@@ -57,8 +57,8 @@ export * from './components/menu'
 export * from './components/dialog'
 export * from './components/avatar'
 export * from './components/skeleton'
-export * from './components/table/tableContainer'
-export * from './components/table/tableToolbar'
+export * from './components/table/table-container'
+export * from './components/table/table-toolbar'
 
 export * from './hooks/use-portal'
 export * from './hooks/use-snack'
